+++ src/translations/de/activeusers.php
@@ -24,5 +24,9 @@ return [
     'User limit' => 'Anzahl Benutzer',
     'Max. shown user' => 'Max. Anzahl der angezeigten Benutzer',
     'Max. content height in px' => 'Max. Höhe des Widget-Inhalts',
-    'A scrollbar is added if the content height exceeds the height set (0 = unlimited)' => 'Ein Scrollbalken wird hinzugefügt, wenn der Inhalt die festgelegte Höhe übersteigt. (0 = unbegrenzt)'
+    'A scrollbar is added if the content height exceeds the height set (0 = unlimited)' => 'Ein Scrollbalken wird hinzugefügt, wenn der Inhalt die festgelegte Höhe übersteigt. (0 = unbegrenzt)',
+    'User link' => 'Verlinkung',
+    'Link for the users. You can use the {{ user }} object' => 'Link zum Benutzer. Du kannst das {{ user }} Objekt benutzen.',
+    'Link target' => 'Link Ziel',
+    'ex. _blank for new window' => 'z.B. _blank für ein neues Fenster'
 ];